Require `caps-lock` key to be held to trigger debug events on macOS
macOS function keys by default trigger system events
 (such as volume, brightness, and keyboard backlight changes) by default, and
 only act as function keys in most programs if the `fn` key is held.
Etterna operates at a low enough level that this does not apply.

This addition explicitly checks in the handler for debug input whether the
 caps-lock key is pressed before doing anything if the platform is macOS.

This is to prevent people from triggering potentially unwanted debug actions when
 trying to change their volume or do other often-used actions.

TODO: Caps-Lock is an incredibly unintuitive key for this behavior.
The best solution to this would be to use the `fn` key's state to determine
 whether events should happen. However, the `fn` key is a special key, and
 can seemingly not be read from the IOHIDDevice layer we're using for input
 on macOS. If there is a way to do it, I've managed to miss it over the hours
 spent on this.
If the input layer is ever rewritten, this should be changed if possible.