Add RageSharedMutex object
This was created to provide the same functionality as a C++ 17 shared_mutex on top of
a RageMutex.
The RageSharedMutex allows an arbitrary number of shared locks or a single exclusive lock to be made.
A shared lock cannot be made if a exclusive lock is held, and vice-versa.
It is intended for use protecting resources where simultaneous reading between threads is ok but
 simultaneous write accesses or read and write accesses is not.